feat: accept comma as decimal separator (Uruguay locale)
- Add parseDecimal() and normalizeDecimalInput() to utils.js
  Both comma and period are accepted; comma is silently converted to period
  before parsing so internal JS math always uses period
- Add showDecimalWarning() to utils.js: shows a small floating tooltip
  near the input when the user types a period, suggesting they use comma
- Add handleDecimalInput() to ui-handlers.js, exposed via window in main.js
- Change table cell inputs from type=number to type=text (inputmode=decimal)
  so the browser does not reject comma on the fly
- Update data-manager.js: use parseDecimal/normalizeDecimalInput in
  updateDefaultError, importCSVFile (also detects semicolon-separated CSVs
  common when locale uses comma as decimal), and getValidData
- Update chart-manager.js: use parseDecimal for all data point parsing
- Update CSS to apply same styling to type=text inputs

+++ js/chart-manager.js
@@ -11,7 +11,7 @@
 import { AppState } from './state.js';
 import { errorBarsPlugin, bullseyePointsPlugin } from './chart-plugins.js';
 import { calculateFit, calculateDerivative, calculateIntegral, getRegressionCoeffs } from './calculations.js';
-import { extractUnit, formatWithUncertainty } from './utils.js';
+import { extractUnit, formatWithUncertainty, parseDecimal } from './utils.js';
 
 /**
  * Inicializa el gráfico de Chart.js
@@ -248,7 +248,7 @@ export function getDataRange() {
             if (dataset.data && dataset.data.length > 0) {
                 dataset.data.forEach(p => {
                     if (p.x !== undefined && p.x !== null) {
-                        const val = parseFloat(p.x);
+                        const val = parseDecimal(p.x);
                         if (!isNaN(val)) {
                             if (val < minX) minX = val;
                             if (val > maxX) maxX = val;
@@ -300,10 +300,10 @@ export function updateChart(animationMode) {
 
     AppState.series.forEach(serie => {
         const validData = serie.data.filter(p => p.x !== '' && p.y !== '').map(p => ({
-            x: parseFloat(p.x),
-            y: parseFloat(p.y),
-            xError: parseFloat(serie.defaultXError || 0),
-            yError: parseFloat(serie.defaultYError || 0)
+            x: parseDecimal(p.x),
+            y: parseDecimal(p.y),
+            xError: parseDecimal(serie.defaultXError || 0),
+            yError: parseDecimal(serie.defaultYError || 0)
         }));
 
         if (validData.length === 0) return;
